chore(base): Move tests in the correct module.
This patch move tests about the `RoomDisplayName` in the newly created
`display_name` module.
