+++ crates/matrix-sdk-base/src/rooms/mod.rs
@@ -147,40 +147,3 @@ pub(crate) enum AccountDataSource {
     #[default]
     Unstable,
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::RoomDisplayName;
-
-    #[test]
-    fn test_room_alias_from_room_display_name_lowercases() {
-        assert_eq!(
-            "roomalias",
-            RoomDisplayName::Named("RoomAlias".to_owned()).to_room_alias_name()
-        );
-    }
-
-    #[test]
-    fn test_room_alias_from_room_display_name_removes_whitespace() {
-        assert_eq!(
-            "room-alias",
-            RoomDisplayName::Named("Room Alias".to_owned()).to_room_alias_name()
-        );
-    }
-
-    #[test]
-    fn test_room_alias_from_room_display_name_removes_non_ascii_symbols() {
-        assert_eq!(
-            "roomalias",
-            RoomDisplayName::Named("Room±Alias√".to_owned()).to_room_alias_name()
-        );
-    }
-
-    #[test]
-    fn test_room_alias_from_room_display_name_removes_invalid_ascii_symbols() {
-        assert_eq!(
-            "roomalias",
-            RoomDisplayName::Named("#Room,{Alias}:".to_owned()).to_room_alias_name()
-        );
-    }
-}
